Organize skeleton components into separate folder for better structure
