+++ client/src/components/ProductDetail.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import ProductSkeleton from "./ProductSkeleton";
+import ProductSkeleton from "./skeletons/ProductSkeleton";
 import bgImage1 from "../assets/premium_photo-1675186049366-64a655f8f537.avif";
 import bgImage2 from "../assets/premium_photo-1677187301660-5e557d9c0724.avif";
 import bgImage3 from "../assets/airpods-max-select-silver-202011.jpeg";
