+++ client/src/components/skeletons/CategoriesSkeleton.tsx
