+++ client/src/components/skeletons/FeaturedSkeleton.tsx
