+++ client/src/components/skeletons/ProductSkeleton.tsx
