+++ client/src/components/skeletons/RecommendedProductsSkeleton.tsx
